requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Caching>=2.3.0,<3.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
zipp==3.23.0
python-dotenv==1.0.1
email-validator==2.1.1
authlib==1.3.0
requests==2.31.0
urllib3<2
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytest==8.3.4
sentry-sdk[flask]==2.18.0
Flask-Migrate==4.0.4
redis>=4.6.0,<5.0
rq>=1.13.0,<3.0
prometheus_client==0.16.0
alembic==1.11.1
pyotp==2.8.0
bleach==6.0.0
pillow>=10.0.0,<11.0.0
pytesseract>=0.3.11,<0.4.0
marshmallow==3.20.1